requirements.txt
aiogram==2.25.2
pytesseract
Pillow
